long

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = {50, 28, 167};
-        int[] sales1 = {150, 328, 67};
-        int[] sales2 = {500, 208, 367};
+        long[] sales = {50, 28, 167};
+        long[] sales1 = {150, 328, 67};
+        long[] sales2 = {500, 208, 367};
 
         SalesManager manager = new SalesManager(sales);
         SalesManager manager1 = new SalesManager(sales1);
